RFC-7009 - check audience of token before revoking
The authorization server first validates the client credentials (in
case of a confidential client) and then verifies whether the token
was issued to the client making the revocation request.  If this
validation fails, the request is refused and the client is informed
of the error by the authorization server as described below.